progressbar: Move to `wibox.widget`
